Оптимизирует login-page
